Caching moved in separate classes from repositories.
