+++ app/src/main/java/ru/brauer/mvp/model/room/IRoomGithubRepositoryCache.kt
@@ -1,0 +1,9 @@
+package ru.brauer.mvp.model.room
+
+import ru.brauer.mvp.model.githubusers.GithubRepository
+import ru.brauer.mvp.model.githubusers.GithubUser
+
+interface IRoomGithubRepositoryCache {
+    fun save(user: GithubUser, repositories: List<GithubRepository>)
+    fun get(user: GithubUser): List<GithubRepository>
+}
